+++ test/c/CMakeLists.txt
@@ -10,6 +10,7 @@ set(dynamics_c_sources
     dynamics_geometry_tests.c
     dynamics_linkage_tests.c
     dynamics_transfer_function_tests.c
+    dynamics_state_space_tests.c
 )
 
 # Build the tests
