+++ test/c/dynamics_state_space_tests.h
@@ -1,0 +1,10 @@
+#ifndef DYNAMICS_STATE_SPACE_TESTS_H_
+#define DYNAMICS_STATE_SPACE_TESTS_H_
+
+#include <stdbool.h>
+
+bool c_test_state_space_initialize();
+bool c_test_state_space_poles_zeros();
+bool c_test_state_space_to_transfer_function();
+
+#endif
