+++ test/c/dynamics_transfer_function_tests.h
@@ -8,6 +8,5 @@ bool c_test_ccf_form();
 bool c_test_ocf_form();
 bool c_test_poles_zeros();
 bool c_test_tf_multiply();
-bool c_test_state_space_initialize();
 
 #endif
